Fix compiler warning.
* test/testskiplist.c
  (skiplist_get_size): Return size_t instead of int.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1902154 13f79535-47bb-0310-9956-ffa450edef68

+++ test/testskiplist.c
@@ -33,7 +33,7 @@
 static apr_pool_t *ptmp = NULL;
 static apr_skiplist *skiplist = NULL;
 
-static int skiplist_get_size(abts_case *tc, apr_skiplist *sl)
+static size_t skiplist_get_size(abts_case *tc, apr_skiplist *sl)
 {
     size_t size = 0;
     apr_skiplistnode *n;
